+++ src/problems/cvrp/operators/cross_exchange.h
@@ -34,6 +34,9 @@ protected:
   bool t_is_normal_valid;
   bool t_is_reverse_valid;
 
+  const Amount source_delivery;
+  const Amount target_delivery;
+
   virtual void compute_gain() override;
 
 public: