+++ src/problems/vrptw/operators/reverse_two_opt.cpp
@@ -35,12 +35,14 @@ ReverseTwoOpt::ReverseTwoOpt(const Input& input,
 bool ReverseTwoOpt::is_valid() {
   return cvrp::ReverseTwoOpt::is_valid() and
          _tw_t_route.is_valid_addition_for_tw(_input,
+                                              _s_delivery,
                                               s_route.rbegin(),
                                               s_route.rbegin() +
                                                 s_route.size() - 1 - s_rank,
                                               0,
                                               t_rank + 1) and
          _tw_s_route.is_valid_addition_for_tw(_input,
+                                              _t_delivery,
                                               t_route.rbegin() +
                                                 t_route.size() - 1 - t_rank,
                                               t_route.rend(),
@@ -55,12 +57,14 @@ void ReverseTwoOpt::apply() {
                      t_route.rend());
 
   _tw_t_route.replace(_input,
+                      _s_delivery,
                       s_route.rbegin(),
                       s_route.rbegin() + s_route.size() - 1 - s_rank,
                       0,
                       t_rank + 1);
 
   _tw_s_route.replace(_input,
+                      _t_delivery,
                       t_job_ranks.begin(),
                       t_job_ranks.end(),
                       s_rank + 1,