+++ src/problems/vrptw/operators/route_split.cpp
@@ -52,8 +52,13 @@ void RouteSplit::apply() {
   // Empty route holding the end of the split.
   auto& end_route = _empty_tw_route_refs[choice.v_end].get();
   assert(end_route.empty());
+  assert(end_route.vehicle_rank == _empty_route_ranks[choice.v_end]);
+
+  const auto end_delivery =
+    _tw_s_route.delivery_in_range(choice.split_rank, _tw_s_route.size());
 
   end_route.replace(_input,
+                    end_delivery,
                     s_route.begin() + choice.split_rank,
                     s_route.end(),
                     0,
@@ -62,8 +67,13 @@ void RouteSplit::apply() {
   // Empty route holding the beginning of the split.
   auto& begin_route = _empty_tw_route_refs[choice.v_begin].get();
   assert(begin_route.empty());
+  assert(begin_route.vehicle_rank == _empty_route_ranks[choice.v_begin]);
+
+  const auto begin_delivery =
+    _tw_s_route.delivery_in_range(0, choice.split_rank);
 
   begin_route.replace(_input,
+                      begin_delivery,
                       s_route.begin(),
                       s_route.begin() + choice.split_rank,
                       0,