+++ src/structures/vroom/break.h
@@ -13,6 +13,7 @@ All rights reserved (see LICENSE).
 #include <string>
 
 #include "structures/typedefs.h"
+#include "structures/vroom/amount.h"
 #include "structures/vroom/time_window.h"
 
 namespace vroom {
@@ -22,13 +23,18 @@ struct Break {
   std::vector<TimeWindow> tws;
   Duration service;
   std::string description;
+  std::optional<Amount> max_load;
 
   Break(Id id,
-        const std::vector<TimeWindow>& tws,
+        const std::vector<TimeWindow>& tws =
+          std::vector<TimeWindow>(1, TimeWindow()),
         UserDuration service = 0,
-        const std::string& description = "");
+        const std::string& description = "",
+        const std::optional<Amount>& max_load = std::optional<Amount>());
 
   bool is_valid_start(Duration time) const;
+
+  bool is_valid_for_load(const Amount& load) const;
 };
 
 } // namespace vroom