+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Support for `max_travel_time` at vehicle level (#273)
 - Support for vehicle fixed costs (#528)
 - Support for cost per hour for vehicles (#732)
+- Support for `max_load` constraint at break level (#786)
 - `RouteSplit` local search operator (#788)
 - Advertise `libvroom` in README and wiki (#42)